1455:
- content manager: meta key description has been extended to a maximum length of 2^16 - 1
- alias: target uri's of aliases aren't any longer stored as html code in the .htaccess file
- contact manager: the input data in the e-mail is now sorted according to the defined field order
- community: inactive users won't get removed when an other users tries to register if no activation timeout has been set
- calendar: the links on the days when displaying the three boxes view are now correct
- shop: storing a now product does work again
- changelog updated

1457:
- renamed /admin to /cadmin

1458:
- added alias /admin which redirects to cadmin

1459:
- changed admin path to /cadmin on the summery page of the web installer

1475:
- contact: placeholders of fields which haven't been set won't be shown anymore within the feedback text
- installer: splitted the sql dump into to dumps (structure and data)
- gallery: every category (not only active ones) gets listed on the confirm page (backend)

git-svn-id: http://svn.contrexxlabs.com/contrexx/trunk@1487 5ce06e5d-2b28-0410-aa6a-b3cd9cf52d8a

+++ lib/FRAMEWORK/User.class.php
@@ -227,7 +227,20 @@ class FWUser
 	{
 		global $objDatabase;
 
-		$objDatabase->Execute("DELETE FROM ".DBPREFIX."access_users WHERE active=0 AND restore_key!='' AND restore_key_time<".time());
+		static $userActivationTimeoutStatus = null;
+
+		if (!isset($userActivationTimeoutStatus)) {
+			$objConfig = $objDatabase->SelectLimit("SELECT 1 FROM `".DBPREFIX."community_config` WHERE `name` = 'user_activation_timeout' AND `status` = 1", 1);
+			if ($objConfig !== false && $objConfig->RecordCount() == 1) {
+				$userActivationTimeoutStatus = true;
+			} else {
+				$userActivationTimeoutStatus = false;
+			}
+		}
+
+		if ($userActivationTimeoutStatus) {
+			$objDatabase->Execute("DELETE FROM ".DBPREFIX."access_users WHERE active=0 AND restore_key!='' AND restore_key_time<".time());
+		}
 	}
 
 	/**
